move back to previous dir structure
