+++ internal/startup/storage.go
@@ -1,10 +1,10 @@
 package startup
 
 import (
-	"github.com/jukeizu/treediagram/pkg/publishing"
-	"github.com/jukeizu/treediagram/pkg/registration"
-	"github.com/jukeizu/treediagram/pkg/scheduling"
-	"github.com/jukeizu/treediagram/pkg/user"
+	"github.com/jukeizu/treediagram/services/publishing"
+	"github.com/jukeizu/treediagram/services/registration"
+	"github.com/jukeizu/treediagram/services/scheduling"
+	"github.com/jukeizu/treediagram/services/user"
 )
 
 type Storage struct {
