+++ listeners/discord/mapper.go
