+++ services/publishing/discord/mapper.go
