+++ services/publishing/logging.go
