+++ services/publishing/publisher.go
