+++ services/registration/endpoints.go
