+++ services/registration/grpc_binding.go
