+++ services/registration/logging.go
